ReportBean fixed

+++ src/main/java/trng/hibernat/beans/ReportBean.java
@@ -11,10 +11,11 @@ public class ReportBean implements Serializable {
 	 * 
 	 */
 	private static final long serialVersionUID = 1L;
-	long customerID;
+	int customerID;
 	String firstName;
 	String lastName;
-	public ReportBean(long customerID, String firstName, String lastName) {
+	
+	public ReportBean(int customerID, String firstName, String lastName) {
 		super();
 		this.customerID = customerID;
 		this.firstName = firstName;
